requirements.txt
streamlit
audio-recorder-streamlit
google-genai